Fix: double factor -> int factor

+++ src/main/java/com/tp/poo/backend/model/behaviour/Divisible.java
@@ -12,8 +12,8 @@ public interface Divisible {
         }
     }
 
-    Set<Figure> vDivision(double factor);
+    Set<Figure> vDivision(int factor);
 
-    Set<Figure> hDivision(double factor);
+    Set<Figure> hDivision(int factor);
     
 }
